requirements.txt
gitpython
chromadb
sentence-transformers
google-generativeai
streamlit
python-dotenv